rework: cleanup CliTester
This factors out the cli tests and adds a few things to make it more robust and support windows
