+++ tox.ini
@@ -34,7 +34,7 @@ passenv =
 extras =
     testing
 commands =
-    pytest -k cli_example --cli-readme-mock=docs/cli-example.md {posargs}
+    pytest -k cli_example --cli-readme-mock=docs/ {posargs}
 
 
 # To run `tox -e lint` you need to make sure you have a
